+++ actix-server/CHANGES.md
@@ -2,11 +2,23 @@
 
 ## Unreleased
 
-- Minimum supported Rust version (MSRV) is now 1.88.
-- internal: Use `core::future::{ready, Ready}` instead of actix-utils'
-- actix-server no longer removes the uds socket [#442]
+## 2.9.1
 
-[#442]: https://github.com/actix/actix-net/pull/442
+- Fix race condition where accept thread shutdown cause cause workers to be dropped prematurely.
+
+## 2.9.0
+
+- Use Tokio `LocalRuntime`s for workers instead of `LocalSet`s.
+
+## 2.8.0
+
+- Remove experimental `io-uring` support.
+
+## 2.7.0
+
+- Add `ServerBuilder::graceful_shutdown_signal()` to propagate graceful shutdown to services.
+- UDS sockets are no longer removed on pause.
+- Minimum supported Rust version (MSRV) is now 1.88.
 
 ## 2.6.0